application/requirements.txt
# Web 框架及其依赖
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==25.1
asgiref==3.8.1
pyecharts==1.9.5
# API 文档
drf-yasg==1.21.8

# 数据库相关
mysql-connector-python==9.2.0
mysqlclient==2.2.7
PyMySQL==1.1.1

# 数据处理
pandas==2.2.3
numpy==2.2.1
python-dateutil==2.9.0.post0

# 数据爬取和解析
selenium==4.27.1
beautifulsoup4==4.12.3
lxml==5.3.0
webdriver-manager==4.0.2

# Excel 处理
openpyxl==3.1.5
xlwt==1.3.0

# 搜索引擎
elasticsearch==8.17.1
elastic-transport==8.17.0

# 工具包
python-dotenv==1.0.1
PyJWT==1.7.1
requests==2.32.3
urllib3==2.3.0
PyYAML==6.0.2